finished delete company

+++ src/main/java/com/excilys/persistence/DAOCompany.java
@@ -98,7 +98,7 @@ public class DAOCompany  {
 		}
 		return res;
 	}
-	public void deleteCompany(int intId) {
+	public void deleteCompany(long intId) {
 		try (PreparedStatement pstmt = Connecticut.conn.prepareStatement(SQLRequests.DELETECOMPANY.getQuery())) {
 			pstmt.setLong(1, intId);
 			pstmt.executeUpdate();
